переделал лаб 2
